Update blend service layer (#75)
A lot of reusable code related to blend processing resides in the api layer itself. Due to it the controllers are being imported by other controllers and worse, even one of the services. Move the reusable code to the service layer.

Ideally every logic should go inside the service layer but that significantly increases the scope. Let's have that as a tech debt for now.