Implement credit based export (#160)
During export, if version is not old nor is the user pro, check with credit service if the user has enough credits as well. Apply watermark only if all 3 are false.

If the reason for entitlement for watermark-free export is credits, for data-sync, save the activity log id. Send the blend id to credit service as well.