Fix deletion retries (#261)
If firebase account was deleted in a deletion request but the request itself failed for some other reason, the retries fail even if the other reason is resolved.

This was already fixed, but for whatever reason, the processors running the queue worker isn't parsing the error handling correctly. Specifically it's not being able to identify the error thrown by the firebase service as "UserError".

For now, remove error type checking. Checking the error code suffices.