Fix race issue in fetching credit service account details (#308)
In the case of a credits account getting created and credits getting added concurrently, we'd previously observed credit addition failing saying an account doesn't exist yet. We "fixed" this by doing a get-or-create. But since dynamodb is an eventually consistent db and especially doesn't support consistent-read with a GSI, which is what we use here for the fetch, we could end up with duplicates.

Now we've observed the same error happening in fetching credits ledger as well.

In both cases, let's simply retry fetching the account a couple of more times based on an exponential backoff strategy.